+++ tests/fold_test.py
@@ -17,7 +17,7 @@ from grscheller.fp.nada import Nada, nada
 
 class Test_fp_folds:
     def test_fold(self) -> None:
-        def add(ii: int, jj: int) -> int:
+        def add2(ii: int, jj: int) -> int:
             return ii+jj
 
         def funcL(acc: int, jj: int) -> int:
@@ -31,47 +31,47 @@ class Test_fp_folds:
         data3: tuple[int, ...] = ()
         data4 = 42,
 
-        assert foldL(data1, add) == 5050
-        assert foldR(data1, add) == 5050
-        assert foldL(data1, add, 10) == 5060
-        assert foldR(data1, add, 10) == 5060
+        assert foldL(data1, add2) == 5050
+        assert foldR(data1, add2) == 5050
+        assert foldL(data1, add2, 10) == 5060
+        assert foldR(data1, add2, 10) == 5060
 
-        assert foldL(data2, add) == 5049
-        assert foldR(data2, add) == 5049
-        assert foldL(data2, add, 10) == 5059
-        assert foldR(data2, add, 10) == 5059
+        assert foldL(data2, add2) == 5049
+        assert foldR(data2, add2) == 5049
+        assert foldL(data2, add2, 10) == 5059
+        assert foldR(data2, add2, 10) == 5059
 
-        assert foldL(data3, add) is nada
-        assert foldR(data3, add) is nada
-        assert foldL(data3, add, 10) == 10
-        assert foldR(data3, add, 10) == 10
+        assert foldL(data3, add2) is nada
+        assert foldR(data3, add2) is nada
+        assert foldL(data3, add2, 10) == 10
+        assert foldR(data3, add2, 10) == 10
 
-        assert foldL(data4, add) == 42
-        assert foldR(data4, add) == 42
-        assert foldL(data4, add, 10) == 52
-        assert foldR(data4, add, 10) == 52
+        assert foldL(data4, add2) == 42
+        assert foldR(data4, add2) == 42
+        assert foldL(data4, add2, 10) == 52
+        assert foldR(data4, add2, 10) == 52
 
         stuff1 = (1, 2, 3, 4, 5)
         stuff2 = (2, 3, 4, 5)
         stuff3: list[int] = []
         stuff4 = 42,
 
-        assert foldL(stuff1, add, default=None) == 15
-        assert foldL(stuff1, add, None, default=None) == 15
-        assert foldL(stuff1, add, 10, default=None) == 25
-        assert foldR(stuff1, add, default=None) == 15
-        assert foldL(stuff2, add, default=None) == 14
-        assert foldR(stuff2, add, default=None) == 14
-        assert foldL(stuff3, add, default=None) == None
-        assert foldR(stuff3, add, default=None) == None
-        assert foldL(stuff4, add, default=None) == 42
-        assert foldR(stuff4, add, default=None) == 42
-        assert foldL(stuff3, add, default=nada) is nada
-        assert foldL(stuff3, add, default=nada) != nada
-        assert foldL(stuff3, add) is nada
-        assert foldR(stuff3, add) is nada
-        assert foldL(stuff4, add) == 42
-        assert foldR(stuff4, add) == 42
+        assert foldL(stuff1, add2) == 15
+        assert foldL(stuff1, add2, None, default=None) == 15
+        assert foldL(stuff1, add2, 10, default=1000) == 25
+        assert foldR(stuff1, add2, default=None) == 15
+        assert foldL(stuff2, add2, default=None) == 14
+        assert foldR(stuff2, add2, default=None) == 14
+        assert foldL(stuff3, add2, default=None) == None
+        assert foldR(stuff3, add2, default=None) == None
+        assert foldL(stuff4, add2, default=None) == 42
+        assert foldR(stuff4, add2, default=None) == 42
+        assert foldL(stuff3, add2, default=nada) is nada
+        assert foldL(stuff3, add2, default=nada) != nada
+        assert foldL(stuff3, add2) is nada
+        assert foldR(stuff3, add2) is nada
+        assert foldL(stuff4, add2) == 42
+        assert foldR(stuff4, add2) == 42
 
         assert foldL(stuff1, funcL, default=None) == -156
         assert foldR(stuff1, funcR, default=None) == 0
@@ -89,46 +89,3 @@ class Test_fp_folds:
         assert foldR(stuff3, funcR) is nada
         assert foldL(stuff4, funcL) == 42
         assert foldR(stuff4, funcR) == 42
-
-    def test_scfold(self) -> None:
-        def add(ii: int|Nada, jj: int|Nada) -> int|Nada:
-            if ii is nada or jj is nada:
-                return -1
-            if (kk := ii+jj) < 42:
-                return kk
-            else:
-                return nada
-
-        data1 = (1, 2, 3, 4, 5, nada, 6, 7, 8, 9, 10)
-        data2 = (1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
-        data3 = [1, 2, 3, 4, 5, 6]
-        data4: tuple[int, ...] = ()
-        data5 = 10,
-        data6 = 15, 20, 25, 30
-
-    #    assert sc_foldL(data1, add, nothing) == 15
-    ##   assert sc_foldR(data1, add, nothing) == 15
-    #    assert sc_foldL(data2, add, nothing) == 36
-    ##   assert sc_foldR(data2, add, nothing) == 36
-    #    assert sc_foldL(data3, add, nothing) == 21
-    ##   assert sc_foldR(data3, add, nothing) == 21
-    #    assert sc_foldL(data4, add, nothing) == nothing
-    ##   assert sc_foldR(data4, add, nothing) == nothing
-    #    assert sc_foldL(data5, add, nothing) == 10
-    ##   assert sc_foldR(data5, add, nothing) == 10
-    #    assert sc_foldL(data6, add, nothing) == 35
-    ##   assert sc_foldR(data6, add, nothing) == 30
-    #    assert sc_foldL(data1, add, nothing, 10) == 25
-    ##   assert sc_foldR(data1, add, nothing, 10) == 25
-    #    assert sc_foldL(data2, add, nothing, 10) == 38
-    ##   assert sc_foldR(data2, add, nothing, 10) == 37
-    #    assert sc_foldL(data3, add, nothing, 20) == 41
-    ##   assert sc_foldR(data3, add, nothing, 20) == 39
-    #    assert sc_foldL(data4, add, nothing, 10) == 10
-    ##   assert sc_foldR(data4, add, nothing, 10) == 10
-    #    assert sc_foldL(data5, add, nothing, 10) == 20
-    ##   assert sc_foldR(data5, add, nothing, 10) == 20
-    #    assert sc_foldL(data6, add, nothing, 10) == 25
-    ##   assert sc_foldR(data6, add, nothing, 10) == 40
-
-